stt_app/requirementss.txt
aniso8601==3.0.2
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Moment==0.6.0
flask-restplus==0.9.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gcloud==0.18.3
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-speech==0.36.0
google-cloud-storage==1.12.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.15.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.5
jsonschema==2.6.0
MarkupSafe==1.0
oauth2client==4.1.3
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pydub==0.23.0
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
